requirements.txt
Django==2.0.6
django-kronos==1.0
django-ipware==2.1.0
